Migrate ChangeEventBuilder into separate file
